Implémentation : Méthode calcul durée opérationnel dans la classe domaine

+++ src/CPP/Trajet.cpp
@@ -6,7 +6,7 @@
 
 //Constructeur
 
-Trajet::Trajet(const int &_num, const std::string &_nom, const std::string &_type, const Sommet *depart, const Sommet *arrivee, float _duree)
+Trajet::Trajet(const int &_num, const std::string &_nom, const std::string &_type, const Sommet *depart, const Sommet *arrivee, int _duree)
       :AbstractIdentite(_num,_nom), m_type(_type), m_duree(_duree)
 {
     m_sommets.first=depart;
